refactor: move rate limiting logic to a separate utility file for better organization and maintainability
